+++ include/zephyr/drivers/w1.h
@@ -112,22 +112,6 @@ __subsystem struct w1_driver_api {
 
 /** @cond INTERNAL_HIDDEN */
 __syscall int w1_change_bus_lock(const struct device *dev, bool lock);
-
-static inline int z_impl_w1_change_bus_lock(const struct device *dev, bool lock)
-{
-	struct w1_master_data *ctrl_data = (struct w1_master_data *)dev->data;
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	if (api->change_bus_lock) {
-		return api->change_bus_lock(dev, lock);
-	}
-
-	if (lock) {
-		return k_mutex_lock(&ctrl_data->bus_lock, K_FOREVER);
-	} else {
-		return k_mutex_unlock(&ctrl_data->bus_lock);
-	}
-}
 /** @endcond */
 
 /**
@@ -190,13 +174,6 @@ static inline int w1_unlock_bus(const struct device *dev)
  */
 __syscall int w1_reset_bus(const struct device *dev);
 
-static inline int z_impl_w1_reset_bus(const struct device *dev)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->reset_bus(dev);
-}
-
 /**
  * @brief Read a single bit from the 1-Wire bus.
  *
@@ -206,13 +183,6 @@ static inline int z_impl_w1_reset_bus(const struct device *dev)
  * @retval -errno Negative error code on error.
  */
 __syscall int w1_read_bit(const struct device *dev);
-
-static inline int z_impl_w1_read_bit(const struct device *dev)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->read_bit(dev);
-}
 
 /**
  * @brief Write a single bit to the 1-Wire bus.
@@ -225,13 +195,6 @@ static inline int z_impl_w1_read_bit(const struct device *dev)
  */
 __syscall int w1_write_bit(const struct device *dev, const bool bit);
 
-static inline int z_impl_w1_write_bit(const struct device *dev, bool bit)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->write_bit(dev, bit);
-}
-
 /**
  * @brief Read a single byte from the 1-Wire bus.
  *
@@ -241,13 +204,6 @@ static inline int z_impl_w1_write_bit(const struct device *dev, bool bit)
  * @retval -errno  Negative error code on error.
  */
 __syscall int w1_read_byte(const struct device *dev);
-
-static inline int z_impl_w1_read_byte(const struct device *dev)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->read_byte(dev);
-}
 
 /**
  * @brief Write a single byte to the 1-Wire bus.
@@ -259,13 +215,6 @@ static inline int z_impl_w1_read_byte(const struct device *dev)
  * @retval -errno Negative error code on error.
  */
 __syscall int w1_write_byte(const struct device *dev, uint8_t byte);
-
-static inline int z_impl_w1_write_byte(const struct device *dev, uint8_t byte)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->write_byte(dev, byte);
-}
 
 /**
  * @brief Read a block of data from the 1-Wire bus.
@@ -302,14 +251,6 @@ __syscall int w1_write_block(const struct device *dev,
  */
 __syscall size_t w1_get_slave_count(const struct device *dev);
 
-static inline size_t z_impl_w1_get_slave_count(const struct device *dev)
-{
-	const struct w1_master_config *ctrl_cfg =
-		(const struct w1_master_config *)dev->config;
-
-	return ctrl_cfg->slave_count;
-}
-
 /**
  * @brief Configure parameters of the 1-Wire master.
  *
@@ -326,14 +267,6 @@ static inline size_t z_impl_w1_get_slave_count(const struct device *dev)
  */
 __syscall int w1_configure(const struct device *dev,
 			   enum w1_settings_type type, uint32_t value);
-
-static inline int z_impl_w1_configure(const struct device *dev,
-				      enum w1_settings_type type, uint32_t value)
-{
-	const struct w1_driver_api *api = (const struct w1_driver_api *)dev->api;
-
-	return api->configure(dev, type, value);
-}
 
 /**
  * @}
@@ -720,6 +653,8 @@ static inline uint16_t w1_crc16(const uint16_t seed, const uint8_t *src,
 /**
  * @}
  */
+
+#include <zephyr/drivers/w1/internal/w1_impl.h>
 #include <zephyr/syscalls/w1.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_W1_H_ */
